HH бот/front_bot/requirements.txt
python-telegram-bot
python-dotenv
httpx
aiohttp